add phase class
