+++ calphy/solid.py
@@ -34,8 +34,9 @@ import pyscal.traj_process as ptp
 from calphy.integrators import *
 import calphy.lattice as pl
 import calphy.helpers as ph
+import calphy.phase as cph
 
-class Solid:
+class Solid(cph.Phase):
     """
     Class for free energy calculation with solid as the reference state
 
@@ -54,84 +55,13 @@ class Solid:
     """
     def __init__(self, options=None, kernel=None, simfolder=None):
 
-        self.options = options
-        self.simfolder = simfolder
-        self.kernel = kernel
+        #call base class
+        super().__init__(options=options,
+        kernel=kernel, simfolder=simfolder)
+
+
         
-        logfile = os.path.join(self.simfolder, "tint.log")
-        self.logger = ph.prepare_log(logfile)
 
-        self.calc = options["calculations"][kernel]
-        self.nsims = self.calc["nsims"]
-
-        self.t = self.calc["temperature"]
-        self.tend = self.calc["temperature_stop"]
-        self.thigh = self.calc["thigh"] 
-        self.p = self.calc["pressure"]
-        self.logger.info("Temperature start: %f K, temperature stop: %f K, pressure: %f bar"%(self.t, self.tend, self.p))
-        
-        if self.calc["iso"]:
-            self.iso = "iso"
-        else:
-            self.iso = "aniso"
-
-        self.l = None
-        self.alat = None
-        self.apc = None
-        self.vol = None
-        self.concentration = None
-        self.prepare_lattice()
-
-        #other properties
-        self.cores = self.options["queue"]["cores"]
-        self.ncells = np.prod(self.calc["repeat"])
-        self.natoms = self.ncells*self.apc        
-        
-        #properties that will be calculated later
-        self.volatom = None
-        self.k = None
-        self.ferr = None
-        self.fref = None
-        self.fideal = None
-        self.w = None
-        self.pv = None
-        self.fe = None
-
-        #box dimensions that need to be stored
-        self.lx = None
-        self.ly = None
-        self.lz = None
-
-        #now manually tune pair styles
-        self.options["md"]["pair_style"] = self.options["md"]["pair_style"][0]
-        self.options["md"]["pair_coeff"] = self.options["md"]["pair_coeff"][0]
-
-    def prepare_lattice(self):
-        """
-        Prepare the lattice for the simulation
-
-        Parameters
-        ----------
-        None
-
-        Returns
-        -------
-        None
-
-        Notes
-        -----
-        Calculates the lattic, lattice constant, number of atoms per unit cell
-        and concentration of the input system.
-        """
-        l, alat, apc, conc = pl.prepare_lattice(self.calc)
-        self.l = l
-        self.alat = alat
-        self.apc = apc
-        self.concentration = conc
-        self.logger.info("Lattice: %s with a=%f"%(self.l, self.alat))
-        self.logger.info("%d atoms in the unit cell"%self.apc)
-        self.logger.info("concentration:")
-        self.logger.info(self.concentration)
 
     def run_averaging(self):
         """
@@ -356,30 +286,6 @@ class Solid:
         self.process_traj()
 
 
-    def process_traj(self):
-        """
-        Process the out trajectory after averaging cycle and 
-        extract a configuration to run integration
-
-        Parameters
-        ----------
-        None
-
-        Returns
-        -------
-        None
-        
-        """
-        trajfile = os.path.join(self.simfolder, "traj.dat")
-        files = ptp.split_trajectory(trajfile)
-        conf = os.path.join(self.simfolder, "conf.dump")
-
-        ph.reset_timestep(files[-1], conf)
-
-        os.remove(trajfile)
-        for file in files:
-            os.remove(file)
-
 
     def run_integration(self, iteration=1):
         """
@@ -538,44 +444,6 @@ class Solid:
         self.fe = self.fref + self.w + self.pv
 
 
-    def submit_report(self):
-        """
-        Submit final report containing results
-
-        Parameters
-        ----------
-        None
-
-        Returns
-        -------
-        None
-        """
-        report = {}
-
-        #input quantities
-        report["input"] = {}
-        report["input"]["temperature"] = int(self.t)
-        report["input"]["pressure"] = float(self.p)
-        report["input"]["lattice"] = str(self.l)
-        report["input"]["element"] = " ".join(np.array(self.options["element"]).astype(str))
-        report["input"]["concentration"] = " ".join(np.array(self.concentration).astype(str))
-
-        #average quantities
-        report["average"] = {}
-        report["average"]["vol/atom"] = float(self.volatom)
-        report["average"]["spring_constant"] = " ".join(np.array(self.k).astype(str))
-        
-        #results
-        report["results"] = {}
-        report["results"]["free_energy"] = float(self.fe)
-        report["results"]["error"] = float(self.ferr)
-        report["results"]["reference_system"] = float(self.fref)
-        report["results"]["work"] = float(self.w)
-        report["results"]["pv"] = float(self.pv)
-
-        reportfile = os.path.join(self.simfolder, "report.yaml")
-        with open(reportfile, 'w') as f:
-            yaml.dump(report, f)
 
 
     def reversible_scaling(self, iteration=1):
@@ -700,26 +568,3 @@ class Solid:
         lmp.close()
 
         
-    def integrate_reversible_scaling(self, scale_energy=False, return_values=False):
-        """
-        Perform integration after reversible scaling
-
-        Parameters
-        ----------
-        scale_energy : bool, optional
-            If True, scale the energy during reversible scaling. 
-
-        return_values : bool, optional
-            If True, return integrated values
-
-        Returns
-        -------
-        res : list of lists of shape 1x3
-            Only returned if `return_values` is True.
-        """
-
-        res = integrate_rs(self.simfolder, self.fe, self.t, self.natoms, p=self.p,
-            nsims=self.nsims, scale_energy=scale_energy, return_values=return_values)
-
-        if return_values:
-            return res
